+++ src/vivarium_inputs/core.py
@@ -202,6 +202,7 @@ def get_relative_risk(entity: Union[Risk, CoverageGap], location_id: int):
         draw_cols = [f'draw_{i}' for i in range(1000)]
         data.loc[:, draw_cols] = 1 / data.loc[:, draw_cols]
         data = handle_gbd_coverage_gap_data(entity, data, 1)
+        data['coverage_gap'] = entity.name
     elif entity.kind == 'coverage_gap':
         data = gbd.get_auxiliary_data('relative_risk', 'coverage_gap', entity.name)
         data['coverage_gap'] = entity.name
@@ -235,10 +236,8 @@ def pull_rr_data_from_gbd(measure_id, location_id):
 def get_population_attributable_fraction(entity, location_id):
     entity_id = get_id_for_measure(entity, 'population_attributable_fraction')
 
-    if entity.kind == 'etiology':
-        data = gbd.get_paf(entity_id=entity_id, location_id=location_id)
-        data = filter_to_most_detailed(data)
-    elif entity.kind == 'risk_factor' and entity.distribution in ['ensemble', 'normal', 'lognormal']:
+    if entity.kind == 'etiology' or \
+            (entity.kind == 'risk_factor' and entity.distribution in ['ensemble', 'normal', 'lognormal']):
         data = gbd.get_paf(entity_id=entity_id, location_id=location_id)
         data = filter_to_most_detailed(data)