+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+**1.0.11 - 11/07/18**
+
+    - Pull pafs by risk instead of by cause.
+
+**1.0.10 - 11/05/18**
+
+ - Make entity/location/measure arguments to functions in core singular.
+ - Fix artifact builder bug causing disappearing builds.
+
 **1.0.9 - 10/26/18**
 
  - Added kind attribute to mapping objects